Rollback rev 19112
git-svn-id: svn://svn.liferay.com/repos/public/portal/trunk@19115 05bdf26c-840f-0410-9ced-eb539d925f36

+++ portal-impl/src/com/liferay/portlet/calendar/service/impl/CalEventLocalServiceImpl.java
@@ -1222,7 +1222,7 @@ public class CalEventLocalServiceImpl extends CalEventLocalServiceBaseImpl {
 
 		int recurrenceType = recurrence.getFrequency();
 
-		int interval = (int)recurrence.getInterval();
+		int interval = recurrence.getInterval();
 
 		if (recurrenceType == Recurrence.DAILY) {
 			recur = new Recur(Recur.DAILY, -1);
@@ -1450,7 +1450,7 @@ public class CalEventLocalServiceImpl extends CalEventLocalServiceBaseImpl {
 				addField = Calendar.YEAR;
 			}
 
-			int addAmount = (int)recurrence.getInterval() * recur.getCount();
+			int addAmount = recurrence.getInterval() * recur.getCount();
 
 			until.add(addField, addAmount);
 			until.add(Calendar.DAY_OF_YEAR, -1);
